+++ Sources/VComponents/Components/Inputs/TextField/VTextField.swift
@@ -177,6 +177,7 @@ public struct VTextField: View {
     @ViewBuilder private var header: some View {
         if let headerTitle = headerTitle, !headerTitle.isEmpty {
             VText(
+                type: .multiLine(alignment: .leading, limit: model.layout.headerLineLimit),
                 color: model.colors.header.for(internalState),
                 font: model.fonts.header,
                 title: headerTitle
@@ -188,7 +189,7 @@ public struct VTextField: View {
     @ViewBuilder private var footer: some View {
         if let footerTitle = footerTitle, !footerTitle.isEmpty {
             VText(
-                type: .multiLine(alignment: .leading, limit: nil),
+                type: .multiLine(alignment: .leading, limit: model.layout.footerLineLimit),
                 color: model.colors.footer.for(internalState),
                 font: model.fonts.footer,
                 title: footerTitle