+++ Sources/VComponents/Components/Item Pickers/Wheel Picker/VWheelPicker.swift
@@ -197,6 +197,7 @@ public struct VWheelPicker<Data, RowContent>: View
     @ViewBuilder private var header: some View {
         if let headerTitle = headerTitle, !headerTitle.isEmpty {
             VText(
+                type: .multiLine(alignment: .leading, limit: model.layout.headerLineLimit),
                 color: model.colors.header.for(internalState),
                 font: model.fonts.header,
                 title: headerTitle
@@ -208,7 +209,7 @@ public struct VWheelPicker<Data, RowContent>: View
     @ViewBuilder private var footer: some View {
         if let footerTitle = footerTitle, !footerTitle.isEmpty {
             VText(
-                type: .multiLine(alignment: .leading, limit: nil),
+                type: .multiLine(alignment: .leading, limit: model.layout.footerLineLimit),
                 color: model.colors.footer.for(internalState),
                 font: model.fonts.footer,
                 title: footerTitle